requirements.txt
pandas
scikit-learn
joblib
python-magic
pdfminer.six
pytesseract
Pillow
libmagic
Flask
Werkzeug
